feat: 삭제하기 전 질문하는 기능 구현

+++ src/components/RestaurantView/handleRestaurantView.ts
@@ -11,6 +11,8 @@ export const onClickCancelButton = () => {
 export const onClickDeleteButton = (id: string) => {
   const deleteButton = document.getElementById("deleteButton");
   deleteButton?.addEventListener("click", () => {
-    deleteRestaurant(id);
+    if (confirm('정말로 삭제하시겠습니까?')) {
+      deleteRestaurant(id);
+    }
   });
 };
